pyproject.toml
[project]
name = "mcp-server-photo-edit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp",
    "mcp[cli]>=1.9.0",
    "Pillow",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]
